+++ src/base/jquery/random.js
@@ -1,0 +1,6 @@
+eventManager.on('jQuery', () => {
+  $.fn.random = function random() {
+    const i = Math.floor(Math.random() * this.length);
+    return $(this[i]);
+  };
+});